+++ application/libraries/Format.php
@@ -461,8 +461,8 @@ class Format {
     // INTERNAL FUNCTIONS
 
     /**
-     * @param $data XML string
-     * @return SimpleXMLElement XML element object; otherwise, empty array
+     * @param string $data XML string
+     * @return array XML element object; otherwise, empty array
      */
     protected function _from_xml($data)
     {
@@ -496,7 +496,7 @@ class Format {
     }
 
     /**
-     * @param $data Encoded json string
+     * @param string $data Encoded json string
      * @return mixed Decoded json string with leading and trailing whitespace removed
      */
     protected function _from_json($data)
@@ -505,7 +505,7 @@ class Format {
     }
 
     /**
-     * @param string Data to unserialized
+     * @param string $data Data to unserialize
      * @return mixed Unserialized data
      */
     protected function _from_serialize($data)
@@ -514,7 +514,7 @@ class Format {
     }
 
     /**
-     * @param $data Data to trim leading and trailing whitespace
+     * @param string $data Data to trim leading and trailing whitespace
      * @return string Data with leading and trailing whitespace removed
      */
     protected function _from_php($data)